+++ packages/tracing/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.2.0 - 2020-06-30
+
+- Added in memory tracer.
+
 ## 0.1.0 - 2020-05-19
 
-- Created the package @fwl/tracing
+- Created the package @fwl/tracing.